templates/data-airflow/skeleton/requirements.txt
# Apache Airflow
apache-airflow>=2.7.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Database connectors
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# API clients
requests>=2.31.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3

# Linting and formatting
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0
mypy>=1.5.0
